requirements.txt
# Core dependencies
python-dotenv>=1.0
pydantic>=2.0
redis>=5.0
qdrant-client>=1.7
fastembed>=0.2

# Agent SDK
claude-agent-sdk>=0.1
anyio>=4.0

# Dashboard API
fastapi>=0.110
uvicorn>=0.27

# Telemetry (optional — set OTEL_EXPORT_MODE=none to skip)
opentelemetry-api>=1.20
opentelemetry-sdk>=1.20
opentelemetry-exporter-otlp-proto-http>=1.20
